Turned form components into modals
