+++ src/components/Modals/index.js
@@ -3,6 +3,8 @@ import DeleteClientModal from './DeleteClientModal';
 import LogoutModal from './LogoutModal';
 import NewMessageModal from './NewMessageModal';
 import UploadDocumentModal from './UploadDocumentModal';
+import SetAclPermissionModal from './SetAclPermissionModal';
+import SetAclPermsDocContainerModal from './SetAclPermsDocContainerModal';
 
 /**
  * Modals related to various functionality within project PASS
@@ -10,4 +12,10 @@ import UploadDocumentModal from './UploadDocumentModal';
  * @namespace Modals
  */
 
-export { AddClientModal, DeleteClientModal, LogoutModal, NewMessageModal, UploadDocumentModal };
+export {
+  AddClientModal,
+  DeleteClientModal,
+  LogoutModal, NewMessageModal,
+  SetAclPermissionsModal,
+  UploadDocumentModal
+};
